Refactor: move beamer files
