--- lib/pages/home_page2.dart
@@ -1,181 +1,0 @@
-import 'package:beamer/beamer.dart';
-import 'package:flutter/material.dart';
-
-// DATA
-const List<Map<String, String>> books = [
-  {
-    'id': '1',
-    'title': 'Stranger in a Strange Land',
-    'author': 'Robert A. Heinlein',
-  },
-  {
-    'id': '2',
-    'title': 'Foundation',
-    'author': 'Isaac Asimov',
-  },
-  {
-    'id': '3',
-    'title': 'Fahrenheit 451',
-    'author': 'Ray Bradbury',
-  },
-];
-
-const List<Map<String, String>> articles = [
-  {
-    'id': '1',
-    'title': 'Explaining Flutter Nav 2.0 and Beamer',
-    'author': 'Toby Lewis',
-  },
-  {
-    'id': '2',
-    'title': 'Flutter Navigator 2.0 for mobile dev: 101',
-    'author': 'Lulupointu',
-  },
-  {
-    'id': '3',
-    'title': 'Flutter: An Easy and Pragmatic Approach to Navigator 2.0',
-    'author': 'Marco Muccinelli',
-  },
-];
-
-// SCREENS
-class BooksScreen extends StatelessWidget {
-  const BooksScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Books'),
-      ),
-      body: ListView(
-        children: books
-            .map(
-              (book) => ListTile(
-                title: Text(book['title']!),
-                subtitle: Text(book['author']!),
-                onTap: () => context.beamToNamed('/books/${book['id']}'),
-              ),
-            )
-            .toList(),
-      ),
-    );
-  }
-}
-
-class BookDetailsScreen extends StatelessWidget {
-  const BookDetailsScreen({required this.book, super.key});
-  final Map<String, String> book;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(book['title']!),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(8),
-        child: Text('Author: ${book['author']}'),
-      ),
-    );
-  }
-}
-
-class ArticlesScreen extends StatelessWidget {
-  const ArticlesScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Articles')),
-      body: ListView(
-        children: articles
-            .map(
-              (article) => ListTile(
-                title: Text(article['title']!),
-                subtitle: Text(article['author']!),
-                onTap: () => context.beamToNamed('/articles/${article['id']}'),
-              ),
-            )
-            .toList(),
-      ),
-    );
-  }
-}
-
-class ArticleDetailsScreen extends StatelessWidget {
-  const ArticleDetailsScreen({required this.article, super.key});
-  final Map<String, String> article;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(article['title']!),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(8),
-        child: Text('Author: ${article['author']}'),
-      ),
-    );
-  }
-}
-
-// LOCATIONS
-class BooksLocation extends BeamLocation<BeamState> {
-  BooksLocation(RouteInformation super.routeInformation);
-
-  @override
-  List<String> get pathPatterns => ['/books/:bookId'];
-
-  @override
-  List<BeamPage> buildPages(BuildContext context, BeamState state) => [
-        const BeamPage(
-          key: ValueKey('books'),
-          title: 'Books',
-          type: BeamPageType.noTransition,
-          child: BooksScreen(),
-        ),
-        if (state.pathParameters.containsKey('bookId'))
-          BeamPage(
-            key: ValueKey('book-${state.pathParameters['bookId']}'),
-            title: books.firstWhere(
-              (book) => book['id'] == state.pathParameters['bookId'],
-            )['title'],
-            child: BookDetailsScreen(
-              book: books.firstWhere(
-                (book) => book['id'] == state.pathParameters['bookId'],
-              ),
-            ),
-          ),
-      ];
-}
-
-class ArticlesLocation extends BeamLocation<BeamState> {
-  ArticlesLocation(RouteInformation super.routeInformation);
-
-  @override
-  List<String> get pathPatterns => ['/articles/:articleId'];
-
-  @override
-  List<BeamPage> buildPages(BuildContext context, BeamState state) => [
-        const BeamPage(
-          key: ValueKey('articles'),
-          title: 'Articles',
-          type: BeamPageType.noTransition,
-          child: ArticlesScreen(),
-        ),
-        if (state.pathParameters.containsKey('articleId'))
-          BeamPage(
-            key: ValueKey('articles-${state.pathParameters['articleId']}'),
-            title: articles.firstWhere(
-              (article) => article['id'] == state.pathParameters['articleId'],
-            )['title'],
-            child: ArticleDetailsScreen(
-              article: articles.firstWhere(
-                (article) => article['id'] == state.pathParameters['articleId'],
-              ),
-            ),
-          ),
-      ];
-}
